Fix InputManager typo in ShadowServiceManager
ShadowServiceManager was using an incorrect reference to InputManager as its
descriptor. It was using the nonexistent android.net.InputManager. This was
likely a copypasta error. Update the descriptor reference to the correct one.

Because the real InputManager is now being used instead of an empty stub, some
additional shadows implementation may be necessary. One issue that came up is
calling InputDevice.getDeviceIds, which calls into InputManager. Add a shadow
for InputManager.getInputDeviceIds to ensure that InputDevice.getDeviceIds does
not crash.

PiperOrigin-RevId: 317133297